Created package util and moved tar management code.
